backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
aiohttp
beautifulsoup4
pydantic==1.10.22
python-multipart
requests
python-dotenv
python-dateutil
sqlmodel
psycopg2-binary
alembic
openai>=1.0.0
passlib[bcrypt]
python-jose[cryptography]
httpx
python-docx
psutil
yandex-cloud-ml-sdk==0.2.0
aiofiles